+++ web/src/refresh-pages/AgentsNavigationPage.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useMemo, useState, useRef, useEffect } from "react";
-import AgentCard from "@/refresh-components/AgentCard";
+import AgentCard from "@/sections/cards/AgentCard";
 import { useUser } from "@/components/user/UserProvider";
 import { checkUserOwnsAssistant as checkUserOwnsAgent } from "@/lib/agents";
 import { useAgents } from "@/hooks/useAgents";
